requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
jinja2==3.1.4
python-multipart>=0.0.9
reportlab>=4.2.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
openai>=1.30.0
psycopg[binary]>=3.2.0
